requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-google-genai
streamlit
langchain-groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
pydantic
altair